lazy: column definitions imports 🍺
