+++ client/components/List/ViewColumnsPanel/hooks/useDragAndDrop.ts
@@ -2,9 +2,9 @@ import { DropResult } from 'react-beautiful-dnd'
 
 /**
  * Hook for handling drag and drop functionality
- * 
+ *
  * @param onReorder - Function to call when items are reordered
- * 
+ *
  * @returns Object containing the drag end handler
  */
 export const useDragAndDrop = (
@@ -12,13 +12,13 @@ export const useDragAndDrop = (
 ) => {
   /**
    * Handle the end of a drag operation
-   * 
+   *
    * @param result - The result of the drag operation
    */
   const handleDragEnd = (result: DropResult) => {
     // If dropped outside the list or no destination, do nothing
     if (!result.destination) return
-    
+
     onReorder(result.source.index, result.destination.index)
   }
