+++ client/pages/Reports/ReportsList/columns/definitions/duration.ts
@@ -1,0 +1,11 @@
+import { TimeEntry } from 'types'
+import { createColumnDef, CreateColumnDefFunction } from 'utils'
+
+/**
+ * Duration column definition for reports list
+ */
+export const durationColumn: CreateColumnDefFunction = (t) =>
+  createColumnDef<TimeEntry>('duration', t('common.durationLabel'), {
+    minWidth: 60,
+    maxWidth: 60
+  })
