+++ client/pages/Reports/ReportsList/columns/definitions/role.name.ts
@@ -1,0 +1,12 @@
+import { TimeEntry } from 'types'
+import { createColumnDef, CreateColumnDefFunction } from 'utils'
+
+/**
+ * Role column definition for reports list
+ */
+export const roleColumn: CreateColumnDefFunction = (t) =>
+  createColumnDef<TimeEntry>('role.name', t('projects.roleFieldLabel'), {
+    minWidth: 100,
+    maxWidth: 140,
+    data: { hidden: true }
+  })
